backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
email-validator==2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
yake==0.4.8
openai==0.28.1
requests==2.31.0
redis==5.0.1
# Rate limiting and security
slowapi==0.1.9
# File processing
PyPDF2==3.0.1
python-docx==1.1.0
# Heavy AI models for enhanced performance
sentence-transformers>=2.2.0
spacy>=3.7.0
# Additional NLP tools
nltk>=3.8.0
transformers>=4.35.0